Turn floats to doubles, skip a bunch of elements to work with nonzeros

+++ headers/helpers.h
@@ -1,11 +1,11 @@
 #ifndef HELPERS_H
 #define HELPERS_H
 
-float calculateDistanceArray(float *p, int start, float *ref, uint dims);
+double calculateDistanceArray(double *p, int start, double *ref, uint dims);
 
-uint partition(float *arr, uint low, uint high);
-float kthSmallest(float *array, uint start, uint end, uint k);
-float quickselect(float *distances, uint end);
+uint partition(double *arr, uint low, uint high);
+double kthSmallest(double *array, uint start, uint end, uint k);
+double quickselect(double *distances, uint end);
 
 
 #endif
